requirements.txt
Flask==3.0.0
gunicorn==22.0.0
twilio==9.0.5
openai==1.51.2
psycopg==3.1.18
gspread==6.1.2
oauth2client==4.1.3
python-dotenv==1.0.1
requests==2.32.3
